madeReportsMapSeparateComponent
