+++ client/src/components/Reports/Reports.tsx
@@ -1,15 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { Report } from '@prisma/client';
-import { GoogleMap } from '@react-google-maps/api';
+import ReportsMap from './ReportsMap';
 
 const Reports: React.FC = () => {
   const [reports, setReports] = useState<Report[]>([]);
-  const [map, setMap] = useState<google.maps.Map>();
-
-  const onLoad = (map: google.maps.Map) => {
-    setMap(map);
-  };
 
   useEffect(() => {
     const fetchReports = async () => {
@@ -29,72 +24,13 @@ const Reports: React.FC = () => {
     fetchReports();
   }, []);
 
-  useEffect(() => {
-    if (map && reports) {
-      const markers = reports.map((report) => {
-        const latLng = { lat: report.location_lat!, lng: report.location_lng! };
-        const marker = new google.maps.Marker({
-          position: latLng,
-          map,
-        });
-
-        const infoWindow = new google.maps.InfoWindow({
-          content: `<div>
-            <p>${report.type}</p>
-            <p>${report.title}</p>
-            <p>${report.body}</p>
-            <p>Reported: ${report.createdAt}</p>
-          </div>`,
-        });
-
-        marker.addListener('click', () => {
-          infoWindow.open(map, marker);
-        });
-
-        return marker;
-      });
-
-      const bounds = new google.maps.LatLngBounds();
-      markers.forEach((marker) => {
-        const position = marker.getPosition();
-        if (position) {
-          bounds.extend(position);
-        }
-      });
-      map.fitBounds(bounds);
-    }
-  }, [map, reports]);
-
-  useEffect(() => {
-    if (map) {
-      if (navigator.geolocation) {
-        navigator.geolocation.getCurrentPosition(
-          (position) => {
-            const pos = {
-              lat: position.coords.latitude,
-              lng: position.coords.longitude,
-            };
-            map.setCenter(pos);
-          },
-          () => {
-            console.error('Error: The Geolocation service failed.');
-          }
-        );
-      } else {
-        console.error("Error: Your browser doesn't support geolocation.");
-      }
-    }
-  }, [map]);
 
   return (
     <div>
-      <GoogleMap
-        mapContainerStyle={{ height: '250px', width: '395px' }}
-        onLoad={onLoad}
-      />
-
+      <ReportsMap />
       <h2>Reports:</h2>
-      {reports.map((report) => (
+      <div id='reports-container'>
+        {reports.map((report) => (
         <div key={report.id}>
           <p>{report.title}</p>
           <p>{report.type}</p>
@@ -103,6 +39,8 @@ const Reports: React.FC = () => {
           </p>
         </div>
       ))}
+      </div>
+
     </div>
   );
 };
