+++ client/src/components/Reports/ReportsMap.tsx
@@ -1,0 +1,102 @@
+import React, { useEffect, useState } from 'react';
+import axios from 'axios';
+import { Report } from '@prisma/client';
+import { GoogleMap } from '@react-google-maps/api';
+
+const ReportsMap: React.FC = () => {
+  const [map, setMap] = useState<google.maps.Map>();
+  const [center, setCenter] = useState<google.maps.LatLng>();
+  const [reports, setReports] = useState<Report[]>([]);
+
+  const onLoad = (map: google.maps.Map) => {
+    setMap(map);
+  };
+
+  useEffect(() => {
+    const fetchReports = async () => {
+      try {
+        const response = await axios.get('/reports');
+        const filteredReports = response.data.filter((report) => {
+          const reportCreatedAt = new Date(report.createdAt);
+          const currentDate = new Date();
+          const monthAgo = new Date(currentDate.getFullYear(), currentDate.getMonth() - 1, currentDate.getDate());
+          return reportCreatedAt >= monthAgo;
+        });
+        setReports(filteredReports);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+    fetchReports();
+  }, []);
+
+  useEffect(() => {
+    if (map && reports) {
+      const markers = reports.map((report) => {
+        const latLng = { lat: report.location_lat!, lng: report.location_lng! };
+        const marker = new google.maps.Marker({
+          position: latLng,
+          map,
+        });
+
+        const infoWindow = new google.maps.InfoWindow({
+          content: `<div>
+            <p>${report.type}</p>
+            <p>${report.title}</p>
+            <p>${report.body}</p>
+            <p>Reported: ${report.createdAt}</p>
+          </div>`,
+        });
+
+        marker.addListener('click', () => {
+          infoWindow.open(map, marker);
+        });
+
+        return marker;
+      });
+
+      const bounds = new google.maps.LatLngBounds();
+      markers.forEach((marker) => {
+        const position = marker.getPosition();
+        if (position) {
+          bounds.extend(position);
+        }
+      });
+      map.fitBounds(bounds);
+    }
+  }, [map, reports]);
+
+  useEffect(() => {
+    if (map) {
+      if (navigator.geolocation) {
+        navigator.geolocation.getCurrentPosition(
+          (position) => {
+            const pos = {
+              lat: position.coords.latitude,
+              lng: position.coords.longitude,
+            };
+            map.setCenter(pos);
+          },
+          () => {
+            console.error('Error: The Geolocation service failed.');
+          }
+        );
+      } else {
+        console.error("Error: Your browser doesn't support geolocation.");
+      }
+    }
+  }, [map]);
+
+  return (
+    <div>
+      <GoogleMap
+        mapContainerStyle={{ height: '250px', width: '395px' }}
+        center={center}
+        zoom={15}
+        onLoad={onLoad}
+      />
+    </div>
+  );
+};
+
+export default ReportsMap;
